backend/requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
python-dotenv==1.0.0
pyrebase4==4.5.0
